New: doc/board-technical.md
This document provides SMBus and related technical details on a
per-board basis.  The source of this information was Supermicro
Technical Support, shared publicly via Email.

I used to have a substantially larger amount of information on many
other boards, and of much higher granularity, but most of those details
were lost in a ZFS-related mishap (when I did not have backups and I was
using a local CVS repository for version control).  I have since
implemented reliable backups and use GitHub.  :-)

I wanted to get the information I do have, from Email archives going all
the way back to February 2009, into GitHub so that there's a public copy
of this information.