CFLAGS: remove -Waggregate-return
This warning is from Days of Olde.  Today's compilers (read: not classic
K&R era) are a bit smarter about handling situations where a struct or
an array could be potentially returned directly on the stack (which is
certainly not what you want with large data).  Actual details of
compiler behaviour and a semi-related discussion are here:

http://www.angelcode.com/dev/callconv/callconv.html
https://bytes.com/topic/c/answers/644271-aggregate-return-warnings